requirement.txt
psycopg2-binary
bcrypt
rich
prompt-toolkit
Pillow
python-dotenv